+++ packages/shared-base/package.json
@@ -12,6 +12,7 @@
   },
   "types": "./dist/index.d.ts",
   "dependencies": {
+    "@dimensiondev/mask-wallet-core": "0.1.0-20211013082857-eb62e5f",
     "@masknet/base": "workspace:^",
     "@masknet/encryption": "workspace:^",
     "@masknet/public-api": "workspace:^",
