+++ packages/shared-base/src/Account/index.ts
@@ -1,0 +1,4 @@
+export interface Account<ChainId> {
+    account: string
+    chainId: ChainId
+}
