+++ packages/web3-shared/base/package.json
@@ -11,7 +11,6 @@
   },
   "types": "./dist/index.d.ts",
   "dependencies": {
-    "@dimensiondev/mask-wallet-core": "0.1.0-20211013082857-eb62e5f",
     "@masknet/public-api": "workspace:^",
     "@masknet/shared-base": "workspace:^",
     "@servie/events": "^3.0.0",
