+++ packages/web3-shared/base/src/specs/index.ts
@@ -11,8 +11,9 @@ import type {
     BindingProof,
     ProfileIdentifier,
     Proof,
+    Account,
+    Wallet,
 } from '@masknet/shared-base'
-import type { api } from '@dimensiondev/mask-wallet-core/proto'
 import type {
     ReturnChainResolver,
     ReturnExplorerResolver,
@@ -811,37 +812,6 @@ export interface AddressName {
     resolvedAddress?: string
 }
 
-export interface Wallet {
-    id: string
-    /** User define wallet name. Default address.prefix(6) */
-    name: string
-    /** The address of wallet */
-    address: string
-    /** true: Mask Wallet, false: External Wallet */
-    hasStoredKeyInfo: boolean
-    /** true: Derivable Wallet. false: UnDerivable Wallet */
-    hasDerivationPath: boolean
-    /** yep: removable, nope: unremovable */
-    configurable?: boolean
-    /** the derivation path when wallet was created */
-    derivationPath?: string
-    /** the derivation path when wallet last was derived */
-    latestDerivationPath?: string
-    /** the internal presentation of mask wallet sdk */
-    storedKeyInfo?: api.IStoredKeyInfo
-    /** the Mask SDK stored key info */
-    /** record created at */
-    createdAt: Date
-    /** record updated at */
-    updatedAt: Date
-    /** an abstract wallet has a owner */
-    owner?: string
-    /** an abstract wallet has been deployed */
-    deployed?: boolean
-    /** persona identifier */
-    identifier?: string
-}
-
 export interface Transaction<ChainId, SchemaType> {
     id: string
     chainId: ChainId
@@ -899,11 +869,6 @@ export interface TokenList<ChainId, SchemaType> {
     tokens: Array<Token<ChainId, SchemaType>>
 }
 
-export interface Account<ChainId> {
-    account: string
-    chainId: ChainId
-}
-
 export interface BalanceEvent<ChainId> {
     /** Emit if the balance of the account updated. */
     update: [Account<ChainId>]
@@ -912,17 +877,6 @@ export interface BalanceEvent<ChainId> {
 export interface BlockNumberEvent<ChainId> {
     /** Emit if the balance of the chain updated. */
     update: [ChainId]
-}
-
-export interface ProviderEvents<ChainId, ProviderType> {
-    /** Emit when the chain id changed. */
-    chainId: [string]
-    /** Emit when the accounts changed. */
-    accounts: [string[]]
-    /** Emit when the site connects with a provider. */
-    connect: [Account<ChainId>]
-    /** Emit when the site disconnect with a provider. */
-    disconnect: [ProviderType]
 }
 
 export interface RecognizableError extends Error {
@@ -934,64 +888,6 @@ export interface WatchEvents<ChainId, Transaction> {
     error: [RecognizableError, JsonRpcPayload]
     /** Emit when the watched transaction status updated. */
     progress: [ChainId, string, TransactionStatusType, Transaction | undefined]
-}
-
-export interface WalletProvider<ChainId, ProviderType, Web3Provider, Web3> {
-    readonly emitter: Emitter<ProviderEvents<ChainId, ProviderType>>
-
-    readonly subscription: {
-        account: Subscription<string>
-        chainId: Subscription<ChainId>
-        wallets: Subscription<Wallet[]>
-    }
-
-    /** Get to know whether the provider is ready. */
-    readonly ready: boolean
-    /** Keep waiting until the provider is ready. */
-    readonly readyPromise: Promise<void>
-    /** connection status */
-    readonly connected: boolean
-    /** async setup tasks */
-    setup(): Promise<void>
-    /** Add a new wallet. */
-    addWallet(wallet: Wallet): Promise<void>
-    /** Update a wallet. */
-    updateWallet(address: string, wallet: Wallet): Promise<void>
-    /** Add or update a new wallet on demand. */
-    updateOrAddWallet(wallet: Wallet): Promise<void>
-    /** Rename a wallet */
-    renameWallet(address: string, name: string): Promise<void>
-    /** Remove a wallet */
-    removeWallet(address: string, password?: string | undefined): Promise<void>
-    /** Update a bunch of wallets. */
-    updateWallets(wallets: Wallet[]): Promise<void>
-    /** Remove a bunch of wallets. */
-    removeWallets(wallets: Wallet[]): Promise<void>
-    /** Switch to the designate account. */
-    switchAccount(account?: string): Promise<void>
-    /** Switch to the designate chain. */
-    switchChain(chainId?: ChainId): Promise<void>
-    /** Create an instance from the network SDK. */
-    createWeb3(options?: ProviderOptions<ChainId>): Web3
-    /** Create an instance that implement the wallet protocol. */
-    createWeb3Provider(options?: ProviderOptions<ChainId>): Web3Provider
-    /** Create the connection. */
-    connect(
-        chainId?: ChainId,
-        address?: string,
-        owner?: {
-            account: string
-            identifier?: ECKeyIdentifier
-        },
-        silent?: boolean,
-    ): Promise<Account<ChainId>>
-    /** Dismiss the connection. */
-    disconnect(): Promise<void>
-}
-
-export interface ProviderOptions<ChainId> {
-    chainId: ChainId
-    account?: string
 }
 
 export interface TransactionChecker<ChainId, Transaction> {
@@ -1647,7 +1543,7 @@ export interface TransactionWatcherState<ChainId, Transaction> {
         status: TransactionStatusType,
     ) => Promise<void>
 }
-export interface ProviderState<ChainId, ProviderType, NetworkType, Web3Provider, Web3> {
+export interface ProviderState<ChainId, ProviderType, NetworkType, WalletProvider> {
     ready: boolean
     readyPromise: Promise<void>
 
@@ -1668,9 +1564,7 @@ export interface ProviderState<ChainId, ProviderType, NetworkType, Web3Provider,
     untilReady: (providerType: ProviderType) => Promise<void>
 
     /** Get a registered wallet provider. */
-    getWalletProvider: (
-        providerType: ProviderType,
-    ) => WalletProvider<ChainId, ProviderType, Web3Provider, Web3> | undefined
+    getWalletProvider: (providerType: ProviderType) => WalletProvider | undefined
 
     /** Connect with the provider and set chain id. */
     connect: (
@@ -1811,6 +1705,7 @@ export interface Web3State<
     TransactionParameter,
     Web3,
     Web3Provider,
+    WalletProvider,
 > {
     AddressBook?: AddressBookState<ChainId>
     BalanceNotifier?: BalanceNotifierState<ChainId>
@@ -1839,7 +1734,7 @@ export interface Web3State<
         Web3,
         Web3Provider
     >
-    Provider?: ProviderState<ChainId, ProviderType, NetworkType, Web3Provider, Web3>
+    Provider?: ProviderState<ChainId, ProviderType, NetworkType, WalletProvider>
     Others?: OthersState<ChainId, SchemaType, ProviderType, NetworkType, Transaction>
     Storage?: Web3StorageServiceState
 }
